fixed move boundary
before the people could move off the grid. we still need to fix them creating more squares. if you look at the print log, 0's are present. we are supposed to remove them. check what the error is. 0 is supposed to indicate that a direction like up or down is invalid. 0's are not being removed however. check line 57.